Trying to get the backend connected to the frontend... lots of changes
